Cleanup and update documentation and scripts

+++ wwwroot/HttpPlatformHandlerStartup.ps1
@@ -1,5 +1,5 @@
 ﻿param(
-    [string]$ApplicationInsightsApiKey = $Env:Deployment_Telemetry_Instrumentation_Key
+    [string]$ApplicationInsightsApiKey = $Env:APPINSIGHTS_INSTRUMENTATIONKEY
 )
 
 function log($message) {
@@ -48,6 +48,26 @@ function TrackEvent {
     }
 }
 
+function Set-Property-Value {
+    param (
+        [string]$ConfigContent, 
+        [string]$PropertyName, 
+        [string]$PropertyValue
+    )
+        
+    if($PropertyName -eq "sonar.jdbc.url") {
+        $dbDriver = $PropertyValue | Select-String -Pattern '^jdbc:([^:]*):' | % {$_.matches.groups[0]}
+        [regex]$pattern = "(?m)^#?$([regex]::Escape($PropertyName))=$([regex]::Escape($dbDriver)).*"
+        $ConfigContent = $pattern.replace($ConfigContent, "$propertyName=$propertyValue", 1)
+    }
+    else {
+        [regex]$pattern = "(?m)^#?$([regex]::Escape($PropertyName))=.*"
+        $ConfigContent = $pattern.replace($ConfigContent, "$propertyName=$propertyValue", 1)
+    }
+
+    return $ConfigContent;
+}
+
 TrackEvent -InstrumentationKey $ApplicationInsightsApiKey -EventName 'Starting HttpPlatformHandler Script'
 
 log('Searching for sonar.properties file')
@@ -57,26 +77,26 @@ if(!$propFile) {
     exit
 }
 log("File found at: $($propFile.FullName)")
-$configContents = Get-Content -Path $propFile.FullName
+$configContents = Get-Content -Path $propFile.FullName -Raw
 
 log('Resetting properties.')
-$configContents = $configContents -ireplace '^#?sonar\.', '#sonar.'
+$configContents = $configContents -ireplace '(?m)^#?sonar\.', '#sonar.'
 
 log('Updating sonar.properties based on environment/application settings.')
 Get-ChildItem Env: | Where-Object -Property Name -like -Value 'sonar.*' | ForEach-Object {
     $propertyName = $_.Name
     $propertyValue = $_.Value
     log("Setting $propertyName to ***VALUE HIDDEN***")
-    $configContents = $configContents -ireplace "^#?$propertyName=.*", "$propertyName=$propertyValue"
+    $configContents = Set-Property-Value $configContents $propertyName $propertyValue
 }
 
-$port = $env:HTTP_PLATFORM_PORT
+$port = $Env:HTTP_PLATFORM_PORT
 log("HTTP_PLATFORM_PORT is: $port")
 log("Updating sonar.web.port to $port")
-$configContents = $configContents -ireplace '^#?sonar\.web\.port=.*', "sonar.web.port=$port"
+$configContents = $configContents -ireplace '(?m)^#?sonar\.web\.port=.*', "sonar.web.port=$port"
 
 log('Saving updated sonar.properties contents')
-$configContents | Out-String | Set-Content -Path $propFile.FullName -NoNewLine
+$configContents.Trim() | Out-String | Set-Content -Path $propFile.FullName -NoNewLine
 
 $sqver = $propFile.FullName.split("\")[4].split("-")[1]
 log("SQ version: $sqver")
@@ -103,14 +123,17 @@ log('Searching for duplicate plugins.')
 $plugins = Get-ChildItem '.\sonarqube-*\extensions\plugins\*' -Filter '*.jar'
 $pluginBaseName = $plugins | ForEach-Object { $_.Name.substring(0, $_.Name.LastIndexOf('-')) }
 $uniquePlugins = $pluginBaseName | Select-Object -Unique
-$duplicates = Compare-Object -ReferenceObject $uniquePlugins -DifferenceObject $pluginBaseName
-if($duplicates)
+if($uniquePlugins)
 {
-    log("Duplicates plugins found for: $($duplicates.InputObject)")a
-    foreach($duplicate in $duplicates) { 
-        $oldestFile = $plugins | Where-Object { $_.Name -imatch $duplicate.InputObject } | Sort-Object -Property 'LastWriteTime' | Select-Object -First 1
-        log("Deleting $oldestFile")
-        $oldestFile | Remove-Item
+    $duplicates = Compare-Object -ReferenceObject $uniquePlugins -DifferenceObject $pluginBaseName
+    if($duplicates)
+    {
+        log("Duplicates plugins found for: $($duplicates.InputObject)")a
+        foreach($duplicate in $duplicates) { 
+            $oldestFile = $plugins | Where-Object { $_.Name -imatch $duplicate.InputObject } | Sort-Object -Property 'LastWriteTime' | Select-Object -First 1
+            log("Deleting $oldestFile")
+            $oldestFile | Remove-Item
+        }
     }
 }
